my_langchain/Langchain/app.py
import os
from dotenv import load_dotenv

from langchain_community.llms import Ollama
import streamlit as st
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser

load_dotenv()


prompt = ChatPromptTemplate.from_template(
    """

    Answer the following question based only on the provided context:
    <context>
    {context}
    </context>

    Keep the provided input strictly in consideration:
    <input>
    {input}
    </input>

    """
)




# Langsmith Tracking
os.environ["LANGCHAIN_API_KEY"] = os.getenv("LANGCHAIN_API_KEY")
os.environ["LANGCHAIN_TRACING_V2"] = "true"
os.environ["LANGCHAIN_PROJECT"] = os.getenv("LANGCHAIN_PROJECT")

