requirements.txt
ipykernel
python-dotenv
langchain
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-community
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openai
beautifulsoup4
streamlit